requirements.txt
charm-crypto>=0.50
pycryptodome>=3.10.1
numpy
pandas
